Fix output detection in global-circulation-model init script
The (first) connected output is supposed to be rotated. As a
consequence, the touch overlay needs to be mapped to the new screen
layout.

`xrandr` adds the `connected` label to each connected display followed
by the resolution and position of the display in the screen layout.
However, additional labels such as `primary` may be added to the line,
which breaks the previous display detection mechanism.

This patch allows for certain additional labels between the `connected`
label and the resolution.